Abstracted argument parsing; win32 fixes
Converting argument parsing to use strtod/strtoul instead of sscanf,
particularly because win32 sscanf doesn't like "inf", but win32 strtod
is fine with it.

Convert output to %.7g instead of %.6f as it's nicer for small values.

Explicitly print "inf" for an infinite result (i.e. the critical value
for a test level of 0) since %f doesn't seem to do that on win32.