Header install location fix; components for win32
The 1.0.0 release binary packages were installing headers in the wrong
place (include/common.hpp, etc. instead of include/fracdist/common.hpp),
and were missing the fracdist/data.hpp header.

Added component configurations for Windows installers so that Windows
users can elect to not install the headers and/or docs when using the
.exe installer.