Added licence note about mn-files data citation
I believe that the upstream citation *requirement* is legally void, as
a program cannot claim copyright over its output (in most cases).
Since the copyright itself only applies to the use (and copying) of the
program itself, any restriction on what you do *after* using the
program is impossible to enforce and, thus, void.

Ethically, however, the citation is required--and thus I interpret
upstream's intentions as a citation request, rather than the
(unenforceable) citation requirement, which *is* compatible with the
GPL.